Conflicts:
	src/main/webapp/resources/CSS/user/mypage.css
	target/classes/com/ltns/rest_area/domain/user/UserDAO.xml
	target/m2e-wtp/web-resources/META-INF/maven/com.ltns/rest_area/pom.properties

+++ src/main/webapp/admin/JS/schedule.js
@@ -1,20 +1,19 @@
 
 
-//date pick  ( jquery ui ) 
-
-
+//crud 
 function click_delete() {
 		var url = "deletePopup";
 		var name = "deletePopup";
-		var option = "width = 600, height = 200 left = 100, top=50,location=no";
+		var option = "width = 800, height = 180 left = 200, top=150,location=no";
 		window.open(url,name,option)
+
 }; 
 
 
 function click_add() {
 	var url = "schedulePopup";
 	var name = "schedulePopup";
-	var option = "width = 600, height = 600 left = 100, top=50,location=no";
+	var option = "width = 600, height = 350 left = 200, top=150,location=no";
 	window.open(url,name,option)
 }; 
 
@@ -22,14 +21,13 @@ function click_add() {
 function click_update() {
 	var url = "updatePopup";
 	var name = "updatePopup";
-	var option = "width = 600, height = 720 left = 100, top=50,location=no";
+	var option = "width = 600, height = 450 left = 200, top=150,location=no";
 	window.open(url,name,option)
 }; 
 
 
 
 $(function () {
-	
 	$.datepicker.setDefaults({
 		dateFormat : 'yy-mm-dd',	//날짜 포맷
 		showOtherMonths : true,		// 다른달 보여주기
@@ -50,6 +48,7 @@ $(function () {
 	
 	$("#startDate").datepicker('setDate', 'today'); //시작일
 	$("#endDate").datepicker('setDate', 'today'); 
+
 	
 	
 	
@@ -79,40 +78,77 @@ $.fn.serializeObject = function(){
 
 function click_ok(){
 	
+	var subject = $('#subject').val();
+	if(subject == null || subject == ""){
+		alert("제목이 비었습니다.");
+		$('#subject').focus();
+		return false; 
+		
+	}
+
+	
+
+	var startDate = $('#startDate').val();
+	var endDate = $('#endDate').val();
+
+	if( (new Date(startDate).getTime() - new Date(endDate).getTime()) > 0){
+		alert("종료일은 시작일 이전으로는 할 수 없습니다.");
+		return false;
+	}
 	var scheduleData = JSON.stringify($('form#scheduleData').serializeObject());
+
 	$.ajax({
 		data : scheduleData,
 		url : "schedule",
 		type : "POST",
 		dataType : "JSON",
 		contentType:'application/json;',
-		success : function(data) {
-			
+		success : function(data,status) {
+			if(data.status == "OK"){
+				alert("일정이 추가되었습니다.")
+				opener.parent.location.reload();
+				location.reload();
+				window.close();
+			}else{
+				alert("에러 발생");
+			}
 		}
 	});
-	alert("일정이 추가되었습니다.")
-	opener.parent.location.reload();
-	location.reload();
-	window.close();
+	
+	
 };
 
 
 function delete_ok(){
+	var subject = $('#subject').val();
+	if(subject == null || subject == ""){
+		alert("제목이 비었습니다.");
+		$('#subject').focus();
+		return false; 
+		
+	}
+
+	
 	var scheduleData = JSON.stringify($('form#deleteForm').serializeObject());
 	$.ajax({
 		data : scheduleData,
 		url : "schedule",
 		type : "DELETE",
 		dataType : "JSON",
-		contentType:'application/json;',
-		success : function(data) {
-			
+		contentType:'application/json; charset=utf-8',
+		success : function(data,status) {
+			if(data.status =="OK"){
+				alert("일정이 삭제되었습니다.")
+				opener.parent.location.reload();
+				location.reload();
+				window.close();
+			}else{
+				alert("삭제할 일정이 없습니다.");
+			}
 		}
 	});
-	alert("일정이 삭제되었습니다..")
-	opener.parent.location.reload();
-	location.reload();
-	window.close();
+
+	
 };
 
 
@@ -120,6 +156,39 @@ function delete_ok(){
 
 
 function update_ok(){
+
+
+	var add_subject = $('#add_subject').val();
+	var subject = $('#subject').val();
+
+	if(add_subject == null || add_subject == ""){
+		alert("기존 일정의 제목이 비었습니다.");
+		$('#add_subject').focus();
+		return false;
+	}
+
+
+	if(subject == null || subject == ""){
+		alert("변경하실 일정의 제목이 비었습니다.");
+		$('#subject').focus();
+		return false; 
+		
+	}
+
+	var startDate = $('#startDate').val();
+	var endDate = $('#endDate').val();
+
+	if( (new Date(startDate).getTime() - new Date(endDate).getTime()) > 0){
+		alert("종료일은 시작일 이전으로는 할 수 없습니다.");
+		return false;
+	}
+
+
+
+
+
+
+
 	var scheduleData = JSON.stringify($('form#updateForm').serializeObject());
 	$.ajax({
 		data : scheduleData,
@@ -127,26 +196,35 @@ function update_ok(){
 		type : "PUT",
 		dataType : "JSON",
 		contentType:'application/json;',
-		success : function(data) {
-			
+		success : function(data,status) {
+				if(data.status == "OK"){
+					alert("일정이 수정되었습니다.")
+					opener.parent.location.reload();
+					location.reload();
+					window.close();
+				}else{
+					alert("기존 일정의 제목이 없습니다.")
+				}
+
 		}
 	});
-	alert("일정이 수정되었습니다..")
-	opener.parent.location.reload();
-	location.reload();
-	window.close();
+	
 };
 
 
 
-  
+$(document).ready(function(){
+	changeColor()
+	
+})
 
+function changeColor(){
+	$(".fc-content").click(function(){
+        alert("test");
+    })
+}
   
   
  
  
  
-
-
-
-
